fix: dev: Add only resolved Nexthops to ECMP routes to avoid traffic to slow-path, when adding new Nexthop to ECMP route.
TG-900

When an route is added, if it is ecmp switchd will pass the routes and nexthop
to plugin, only the resolved ones, preference is first resolved is searched,
and if non resolved fill for the port based nexthops and if non port based
nexthop, fill the unresolved ip nexthops.
For non-ecmp switchd will first check if it is resolved and pass that to plugin
otherwise pass as unresolved with existing logic.

Change-Id: I65a7cc837d6edeb72f5816219895e79646c2224b
Signed-off-by: aslammdkhan <aslam.khan@hpe.com>
(cherry picked from commit d779f46b36bd88e6a0375b6a85592e39e4d054f9)